requirements_dev.txt
pip==8.1.2
bumpversion==0.5.3
wheel==0.29.0
watchdog==0.8.3
flake8==2.6.0
tox==2.3.1
coverage==4.1
pytest-aiohttp==0.1.3
Sphinx==1.4.8
cryptography==1.4
PyYAML==3.11
pytest==2.9.2
